+++ src/expressionparser/parser.cpp
@@ -170,7 +170,7 @@ std::string CleanNumericText(std::string_view text)
     return clean_str;
 }
 */
-HRESULT GenerateExecutionSteps(TSNode node, const std::string &source, std::list<ExecutionStep> &program, std::string &output)
+HRESULT GenerateExecutionSteps(TSNode node, const std::string &source, std::list<ExecutionStepData> &program, std::string &output)
 {
     if (ts_node_is_null(node) ||
         // Skip zero-width recovery nodes that aren't explicit errors
@@ -375,8 +375,8 @@ HRESULT GenerateExecutionSteps(TSNode node, const std::string &source, std::list
         // 1. Evaluate the base object first
         IfFailRet(GenerateExecutionSteps(baseExpr, source, program, output));
 
-        // FIXME binding now works with null check all the time. Emit the conditional jump or
-        //       provide as marker in MemberBindingExpression and ElementBindingExpression
+        // FIXME: Binding now works with null check all the time. Emit the conditional jump or
+        //        provide as marker in MemberBindingExpression and ElementBindingExpression.
 
         // 2. Evaluate the binding (e.g., fetching the member or the element)
         IfFailRet(GenerateExecutionSteps(binding, source, program, output));
@@ -789,7 +789,7 @@ HRESULT GenerateExecutionSteps(TSNode node, const std::string &source, std::list
 
 } // unnamed namespace
 
-HRESULT GenerateStackMachineProgram(const std::string &expression, std::list<ExecutionStep> &program, std::string &output)
+HRESULT GenerateStackMachineProgram(const std::string &expression, std::list<ExecutionStepData> &program, std::string &output)
 {
     TSParser *parser = ts_parser_new();
     ts_parser_set_language(parser, tree_sitter_c_sharp());